+++ src/RawRecord/index.d.ts
@@ -1,0 +1,22 @@
+declare module '@nozbe/watermelondb/RawRecord' {
+  import { ColumnName, ColumnSchema, RecordId, TableSchema } from '@nozbe/watermelondb'
+  import { SyncStatus } from '@nozbe/watermelondb/Model'
+
+  export type DirtyRaw = object
+
+  export interface RawRecord {
+    id: RecordId
+    _status: SyncStatus
+    _changed: string
+    last_modified: number | null
+  }
+
+  export function sanitizedRaw(dirtyRaw: DirtyRaw, tableSchema: TableSchema): RawRecord
+
+  export function setRawSanitized(
+    rawRecord: RawRecord,
+    columnName: ColumnName,
+    value: any,
+    columnSchema: ColumnSchema,
+  ): void
+}