+++ src/decorators/action/index.d.ts
@@ -1,0 +1,4 @@
+declare module '@nozbe/watermelondb/decorators/action' {
+  const action: MethodDecorator
+  export default action
+}