+++ src/decorators/common.d.ts
@@ -1,0 +1,10 @@
+declare module '@nozbe/watermelondb/decorators/common' {
+  import { ColumnName } from '@nozbe/watermelondb'
+
+  export function ensureDecoratorUsedProperly(
+    columnName: ColumnName,
+    target: Object,
+    key: string,
+    descriptor: Object,
+  ): void
+}