+++ src/decorators/date/index.d.ts
@@ -1,0 +1,7 @@
+declare module '@nozbe/watermelondb/decorators/date' {
+  import { Decorator, RawDecorator } from '@nozbe/watermelondb/utils/common/makeDecorator'
+  import { ColumnName } from '@nozbe/watermelondb'
+
+  const date: Decorator<[ColumnName], (columnName: ColumnName) => RawDecorator>
+  export default date
+}