+++ src/decorators/immutableRelation/index.d.ts
@@ -1,0 +1,11 @@
+declare module '@nozbe/watermelondb/decorators/immutableRelation' {
+  import { ColumnName, TableName } from '@nozbe/watermelondb'
+  import { Decorator, RawDecorator } from '@nozbe/watermelondb/utils/common/makeDecorator'
+
+  const immutableRelation: Decorator<
+    [TableName<any>, ColumnName],
+    (relationTable: TableName<any>, relationIdColumn: ColumnName) => RawDecorator
+  >
+
+  export default immutableRelation
+}